separate user code from the generated main.c/.h
