+++ Core/Inc/main.h
@@ -31,21 +31,14 @@ extern "C" {
 
 /* Private includes ----------------------------------------------------------*/
 /* USER CODE BEGIN Includes */
-#include "board.h"
-#include <stdio.h>
-#include <stdint.h>
-#include <stdbool.h>
-#include <stdlib.h>
-#include <string.h>
-#include <math.h>
-#include <malloc.h>
+
+#include "user_main.h"
 
 /* USER CODE END Includes */
 
 /* Exported types ------------------------------------------------------------*/
 /* USER CODE BEGIN ET */
-extern IWDG_HandleTypeDef hiwdg;
-extern CRC_HandleTypeDef hcrc;
+
 /* USER CODE END ET */
 
 /* Exported constants --------------------------------------------------------*/
@@ -55,14 +48,6 @@ extern CRC_HandleTypeDef hcrc;
 // Don't uncomment!!
 //               __attribute__((optimize("O0")))
 
-//#define DISABLE_OUTPUT                // Enable to fully disable the tip power (Ex. for debugging)
-#define DEBUG_ERROR                   // Enable to show file/line error messages
-
-#ifdef DEBUG
-extern struct mallinfo mi;
-#define DEBUG_ALLOC                   // Enable alloc debugging in debug builds
-#endif
-
 /* USER CODE END EC */
 
 /* Exported macro ------------------------------------------------------------*/
@@ -71,49 +56,25 @@ extern struct mallinfo mi;
 // STM32CUBE IDE removed line printing in Error Handler. This macro restores it.
 // Credits: https://community.st.com/s/question/0D50X00009XkffVSAR/stm32cubemx-v421-errorhandler-definition-issues-in-mainh
 #ifdef DEBUG_ERROR
-	#define GET_MACRO( _0, _1, NAME, ... ) NAME
-	#define Error_Handler(...) GET_MACRO( _0, ##__VA_ARGS__, Error_Handler1, Error_Handler0 )()
-	#define Error_Handler0() _Error_Handler(__BASE_FILE__, __LINE__ )
-	#define Error_Handler1(unused) _Error_Handler( char * file, int line )
-	void _Error_Handler(char *, int);
+    #define GET_MACRO( _0, _1, NAME, ... ) NAME
+    #define Error_Handler(...) GET_MACRO( _0, ##__VA_ARGS__, Error_Handler1, Error_Handler0 )()
+    #define Error_Handler0() _Error_Handler(__BASE_FILE__, __LINE__ )
+    #define Error_Handler1(unused) _Error_Handler( char * file, int line )
+    void _Error_Handler(char *, int);
 #endif
+
 /* USER CODE END EM */
 
 /* Exported functions prototypes ---------------------------------------------*/
 void Error_Handler(void);
 
 /* USER CODE BEGIN EFP */
-void Program_Handler(void);
 
-
-
-/*
- * Macro to enable debugging of the allocated memory
- * max_allocated will hold the max used memory at any time
- *
- */
-#ifdef DEBUG_ALLOC
-extern uint32_t max_allocated;
-#define dbg_mem() mi=mallinfo();                  \
-                  if(mi.uordblks>max_allocated){  \
-                    max_allocated=mi.uordblks;     \
-                  }                               \
-
-#define _malloc(x)    malloc(x); dbg_mem()
-#define _calloc(x,y)  calloc(x,y); dbg_mem()
-#define _free(x)      free(x); dbg_mem()
-
-#else
-#define _malloc     malloc
-#define _calloc     calloc
-#define _free       free
-#endif
 /* USER CODE END EFP */
 
 /* Private defines -----------------------------------------------------------*/
 /* USER CODE BEGIN Private defines */
-#define WAKE_input()		HAL_GPIO_ReadPin(WAKE_GPIO_Port, WAKE_Pin)
-#define BUTTON_input()		HAL_GPIO_ReadPin(ENC_SW_GPIO_Port, ENC_SW_Pin)
+
 /* USER CODE END Private defines */
 
 #ifdef __cplusplus
@@ -121,5 +82,3 @@ extern uint32_t max_allocated;
 #endif
 
 #endif /* __MAIN_H */
-
-/************************ (C) COPYRIGHT STMicroelectronics *****END OF FILE****/
